requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.0.1
openpyxl==3.1.2
werkzeug==2.3.4
Flask-WTF==1.1.1
Flask-Assets==2.0
cssmin==0.2.0
jsmin==3.0.1
Flask-Caching==2.0.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.15
gunicorn==20.1.0
Flask-Cors==4.0.0
